[feat] 내부 아이템 이동 로직을 short로 변경 완료

+++ Assets/Scripts/Extension/Interface.cs
@@ -47,9 +47,9 @@ public interface ITransporter
 {
     public bool CanItemOut(ITransporter nextNode);
     public void ItemOut(ITransporter nextNode);
-    public bool CanItemIn(int itemId);
-    public void ItemIn(int itemId, Vector3 inPos);
-    public int GetItem();
+    public bool CanItemIn(short itemId);
+    public void ItemIn(short itemId, Vector3 inPos);
+    public short GetItem();
     public void LogicInit();
     public void ExcuteLogic_OnUpdate(float deltaTime);
 }
